Change Sphere center to use Point not Coords and Offsets

+++ src/main/java/com/irtimaled/bbor/common/messages/BoundingBoxSerializer.java
@@ -31,7 +31,7 @@ class BoundingBoxSerializer {
     private static void serializeVillage(BoundingBoxVillage boundingBox, PayloadBuilder builder) {
         builder.writeChar('V')
                 .writeCoords(boundingBox.getCenter())
-                .writeVarInt(boundingBox.getRadius())
+                .writeVarInt((int)boundingBox.getRadius())
                 .writeBoolean(boundingBox.getSpawnsIronGolems())
                 .writeVarInt(boundingBox.getColor().getRGB());
         for (Coords door : boundingBox.getDoors()) {
